Added tests for backend.

+++ backend/src/main/java/com/school/final_project/StoreItem.java
@@ -59,7 +59,7 @@ public class StoreItem {
         this.itemName = itemName;
     }
 
-    public void setItemPrice(float itemPrice) {
+    public void setItemPrice(double itemPrice) {
         this.itemPrice = itemPrice;
     }
